+++ build-pkg/build_n_check.R
@@ -26,6 +26,10 @@ document()
 pkg_loc <- build(path = "build-pkg", args = "--resave-data")
 check_built(path = pkg_loc)
 
+build_manual(path = "build-pkg")
+install()
+
+
 
 # See at
 ## browseURL(paste0("https://www.marinedatascience.co/blog/2020/01/09/",